+++ packages/capabilities/core/index.ts
@@ -1,0 +1,3 @@
+export { defineCapability, type CapabilityDefinition, type CapabilityInput, type CapabilityInputSchema, type CapabilityOutput, type CapabilityOutputSchema } from './definition.ts'
+export { CapabilityExecutionError, type CapabilityExecutionErrorOptions } from './errors.ts'
+export type { CapabilityProvider } from './provider.ts'